fix build and update data structure

+++ source/GSPdbPacker.cpp
@@ -23,7 +23,7 @@ bool GSPdbPacker::WriteTo(const char * pFileName)
         return false;
     }
     // Prepare
-    uint32_t recordCount = m_records.size();
+    uint32_t recordCount = (uint32_t)m_records.size();
     m_header.numRecords = recordCount;
     m_header.uniqueIDSeed = recordCount * 2 + 1;
     GSPdbRecord record;
@@ -35,7 +35,7 @@ bool GSPdbPacker::WriteTo(const char * pFileName)
     {
         record.uniqueID = i * 2;
         record.WriteTo(bytes);
-        record.dataOff += m_records[i].size();
+        record.dataOff += (uint32_t)m_records[i].size();
     }
     GSPushU16BE(bytes, 0); // padding
     for (uint32_t i = 0; i < recordCount; ++i)
